rename: 패키지 구조 리팩토링을 위한 user 관련 클래스 이동 - servlet.user 패키지로 이동
